+++ CHANGES.md
@@ -8,6 +8,7 @@ v?.?.?                                                        (2023-??-?? UTC+8)
   * Flag "-temp" changed to "-tmps".
   * procmain.Main() now accepts a setting parameter.
   * Action "test" is not integrated.
+  * Values can now be variables in config file.
 
 v0.1.7                                                        (2023-06-13 UTC+8)
 ================================================================================